+++ README.md
@@ -4,3 +4,6 @@
 - Title 추가
 - 개인정보 입력 필드 추가
 - 회원가입 버튼 추가
+
+## step3
+- 개인정보 입력 필드 유효성 검사
