+++ app/src/main/java/nextstep/signup/MainActivity.kt
@@ -1,37 +1,14 @@
 package nextstep.signup
 
-import android.annotation.SuppressLint
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import nextstep.signup.ui.theme.SignupTheme
 
 class MainActivity : ComponentActivity() {
@@ -50,92 +27,6 @@ class MainActivity : ComponentActivity() {
     }
 }
 
-@SuppressLint("ResourceAsColor")
-@Composable
-fun SignUpScreen(
-    modifier: Modifier = Modifier
-) {
-    var userName by remember { mutableStateOf("") }
-    var email by remember { mutableStateOf("") }
-    var password by remember { mutableStateOf("") }
-    var passwordConfirm by remember { mutableStateOf("") }
-
-    Column(
-        modifier = modifier.fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 40.dp),
-            text = stringResource(id = R.string.title),
-            fontSize = 26.sp,
-            fontWeight = FontWeight.Bold,
-            textAlign = TextAlign.Center
-        )
-
-        SignUpInput(
-            value = userName,
-            onValueChange = { userName = it },
-            label = stringResource(id = R.string.userNameLabel)
-        )
-
-        SignUpInput(
-            value = email,
-            onValueChange = { email = it },
-            label = stringResource(id = R.string.emailLabel)
-        )
-
-        SignUpInput(
-            value = password,
-            onValueChange = { password = it },
-            visualTransformation = PasswordVisualTransformation(),
-            label = stringResource(id = R.string.passwordLabel)
-        )
-
-        SignUpInput(
-            value = passwordConfirm,
-            onValueChange = { passwordConfirm = it },
-            visualTransformation = PasswordVisualTransformation(),
-            label = stringResource(id = R.string.passwordConfirmLabel)
-        )
-
-        Button(
-            onClick = { /*TODO*/ },
-            colors = ButtonDefaults.buttonColors(
-                disabledContainerColor = Color(R.color.blue),
-                disabledContentColor = Color.White
-            ),
-            content = {
-                Text(
-                    text = stringResource(id = R.string.signUp)
-                )
-            },
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(100.dp)
-                .padding(20.dp)
-        )
-    }
-}
-
-@Composable
-fun SignUpInput(
-    value: String,
-    onValueChange: (String) -> Unit,
-    label: String,
-    visualTransformation: VisualTransformation = VisualTransformation.None,
-) {
-    TextField(
-        value = value,
-        onValueChange = onValueChange,
-        visualTransformation = visualTransformation,
-        label = { Text(text = label) },
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(20.dp)
-    )
-}
 
 @Preview
 @Composable
